# integrate blockwise fp8 kernel to optimize DeepSeekV3