<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->
Integrate #3267 into python side to optimize deepseekv3, merge it after #3267 has merged and sgl kernel is released
## Modifications

<!-- Describe the changes made in this PR. -->

## Test Method
Following #3486, test on gsm8k and mmlu
|    | latency | accuracy |
|  ----  | ----  | --- |
| gsm8k (before) | 97.148s | 0.953 |
| gsm8k (after) | 82.708s | 0.958 |
| mmlu (before) | 250.839s | 0.871|
| mmlu (after) | 240.800s | 0.871|

## TODO
- [ ] remove some of sgl kenel files once CUTLASS has solve this issue https://github.com/NVIDIA/cutlass/issues/2087
- [ ] maybe we can replace `per_token_group_quant_fp8` with `sgl_per_token_group_quant_fp8`, but we need column major scale output.
- [ ] use more robust shape check (not only for 128 x 128 blockwise)

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.